x/program: delete vm package quote> vm package is removed. Few stuff has been moved into internal package
