+++ x/programs/cmd/simulator/cmd/program.go
@@ -17,7 +17,7 @@ import (
 	"github.com/ava-labs/hypersdk/state"
 	hutils "github.com/ava-labs/hypersdk/utils"
 
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/actions"
+	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/internal/actions"
 )
 
 func newProgramCmd(log logging.Logger, db *state.SimpleMutable) *cobra.Command {
