+++ x/programs/cmd/simulator/internal/actions/actions.go
@@ -1,0 +1,6 @@
+package actions
+
+// CallParam defines a value to be passed to a guest function.
+type CallParam struct {
+	Value interface{} `json,yaml:"value"`
+}
