+++ x/programs/cmd/simulator/internal/actions/output.go
@@ -1,0 +1,5 @@
+package actions
+
+var (
+	OutputValueZero = []byte("value is zero")
+)
