+++ x/programs/cmd/simulator/internal/consts/consts.go
@@ -1,6 +1,8 @@
 // Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package actions
+package consts
 
-var OutputValueZero = []byte("value is zero")
+const (
+	HRP = "matrix"
+)
