--- x/programs/cmd/simulator/vm/controller/controller.go
@@ -1,74 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package controller
-
-import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/database"
-	"github.com/ava-labs/avalanchego/snow"
-	"github.com/ava-labs/hypersdk/chain"
-
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/config"
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/genesis"
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/storage"
-)
-
-type Controller struct {
-	snowCtx      *snow.Context
-	genesis      *genesis.Genesis
-	config       *config.Config
-	stateManager *storage.StateManager
-
-	metrics *metrics
-
-	metaDB database.Database
-}
-
-func (c *Controller) Rules(t int64) chain.Rules {
-	// TODO: extend with [UpgradeBytes]
-	return c.genesis.Rules(t, c.snowCtx.NetworkID, c.snowCtx.ChainID)
-}
-
-func (c *Controller) StateManager() chain.StateManager {
-	return c.stateManager
-}
-
-func (c *Controller) Accepted(ctx context.Context, blk *chain.StatelessBlock) error {
-	batch := c.metaDB.NewBatch()
-	defer batch.Reset()
-
-	results := blk.Results()
-	for i, tx := range blk.Txs {
-		result := results[i]
-		if c.config.GetStoreTransactions() {
-			err := storage.StoreTransaction(
-				ctx,
-				batch,
-				tx.ID(),
-				blk.GetTimestamp(),
-				result.Success,
-				result.Consumed,
-				result.Fee,
-			)
-			if err != nil {
-				return err
-			}
-		}
-		if result.Success {
-			// todo manage metrics?
-		}
-	}
-	return batch.Write()
-}
-
-func (*Controller) Rejected(context.Context, *chain.StatelessBlock) error {
-	return nil
-}
-
-func (*Controller) Shutdown(context.Context) error {
-	// Do not close any databases provided during initialization. The VM will
-	// close any databases your provided.
-	return nil
-}
