--- x/programs/cmd/simulator/vm/controller/resolutions.go
@@ -1,33 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package controller
-
-import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/hypersdk/fees"
-
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/genesis"
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/storage"
-)
-
-func (c *Controller) Genesis() *genesis.Genesis {
-	return c.genesis
-}
-
-//func (c *Controller) Logger() logging.Logger {
-//	return c.inner.Logger()
-//}
-
-//func (c *Controller) Tracer() trace.Tracer {
-//	return c.inner.Tracer()
-//}
-
-func (c *Controller) GetTransaction(
-	ctx context.Context,
-	txID ids.ID,
-) (bool, int64, bool, fees.Dimensions, uint64, error) {
-	return storage.GetTransaction(ctx, c.metaDB, txID)
-}
