--- x/programs/cmd/simulator/vm/genesis/errors.go
@@ -1,11 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package genesis
-
-import "errors"
-
-var (
-	ErrInvalidHRP    = errors.New("invalid HRP")
-	ErrInvalidTarget = errors.New("invalid target")
-)
