--- x/programs/cmd/simulator/vm/genesis/rules.go
@@ -1,112 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package genesis
-
-import (
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/hypersdk/chain"
-	"github.com/ava-labs/hypersdk/fees"
-)
-
-var _ chain.Rules = (*Rules)(nil)
-
-type Rules struct {
-	g *Genesis
-
-	networkID uint32
-	chainID   ids.ID
-}
-
-// TODO: use upgradeBytes
-func (g *Genesis) Rules(_ int64, networkID uint32, chainID ids.ID) *Rules {
-	return &Rules{g, networkID, chainID}
-}
-
-func (r *Rules) GetSponsorStateKeysMaxChunks() []uint16 {
-	return r.g.SponsorStateKeysMaxChunks
-}
-
-func (r *Rules) GetMinBlockGap() int64 {
-	return r.g.MinBlockGap
-}
-
-func (r *Rules) GetMinEmptyBlockGap() int64 {
-	return r.g.MinEmptyBlockGap
-}
-
-func (r *Rules) GetValidityWindow() int64 {
-	return r.g.ValidityWindow
-}
-
-func (r *Rules) GetMinUnitPrice() fees.Dimensions {
-	return r.g.MinUnitPrice
-}
-
-func (r *Rules) GetUnitPriceChangeDenominator() fees.Dimensions {
-	return r.g.UnitPriceChangeDenominator
-}
-
-func (r *Rules) GetWindowTargetUnits() fees.Dimensions {
-	return r.g.WindowTargetUnits
-}
-
-func (r *Rules) GetMaxBlockUnits() fees.Dimensions {
-	return r.g.MaxBlockUnits
-}
-
-func (r *Rules) GetBaseComputeUnits() uint64 {
-	return r.g.BaseComputeUnits
-}
-
-func (r *Rules) GetBaseWarpComputeUnits() uint64 {
-	return r.g.BaseWarpComputeUnits
-}
-
-func (r *Rules) GetWarpComputeUnitsPerSigner() uint64 {
-	return r.g.WarpComputeUnitsPerSigner
-}
-
-func (r *Rules) GetOutgoingWarpComputeUnits() uint64 {
-	return r.g.OutgoingWarpComputeUnits
-}
-
-func (r *Rules) GetStorageKeyReadUnits() uint64 {
-	return r.g.StorageKeyReadUnits
-}
-
-func (r *Rules) GetStorageValueReadUnits() uint64 {
-	return r.g.StorageValueReadUnits
-}
-
-func (r *Rules) GetStorageKeyAllocateUnits() uint64 {
-	return r.g.StorageKeyAllocateUnits
-}
-
-func (r *Rules) GetStorageValueAllocateUnits() uint64 {
-	return r.g.StorageValueAllocateUnits
-}
-
-func (r *Rules) GetStorageKeyWriteUnits() uint64 {
-	return r.g.StorageKeyWriteUnits
-}
-
-func (r *Rules) GetStorageValueWriteUnits() uint64 {
-	return r.g.StorageValueWriteUnits
-}
-
-func (*Rules) GetWarpConfig(ids.ID) (bool, uint64, uint64) {
-	return false, 0, 0
-}
-
-func (r *Rules) NetworkID() uint32 {
-	return r.networkID
-}
-
-func (r *Rules) ChainID() ids.ID {
-	return r.chainID
-}
-
-func (*Rules) FetchCustom(string) (any, bool) {
-	return nil, false
-}
