--- x/programs/cmd/simulator/vm/rpc/consts.go
@@ -1,6 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package rpc
-
-const JSONRPCEndpoint = "/simulatorapi"
