--- x/programs/cmd/simulator/vm/rpc/dependencies.go
@@ -1,20 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package rpc
-
-import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/trace"
-	"github.com/ava-labs/hypersdk/fees"
-
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/genesis"
-)
-
-type Controller interface {
-	Genesis() *genesis.Genesis
-	Tracer() trace.Tracer
-	GetTransaction(context.Context, ids.ID) (bool, int64, bool, fees.Dimensions, uint64, error)
-}
