--- x/programs/cmd/simulator/vm/rpc/errors.go
@@ -1,8 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package rpc
-
-import "errors"
-
-var ErrTxNotFound = errors.New("tx not found")
