--- x/programs/cmd/simulator/vm/rpc/jsonrpc_server.go
@@ -1,60 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package rpc
-
-import (
-	"net/http"
-
-	"github.com/ava-labs/avalanchego/ids"
-
-	"github.com/ava-labs/hypersdk/fees"
-
-	"github.com/ava-labs/hypersdk/x/programs/cmd/simulator/vm/genesis"
-)
-
-type JSONRPCServer struct {
-	c Controller
-}
-
-func NewJSONRPCServer(c Controller) *JSONRPCServer {
-	return &JSONRPCServer{c}
-}
-
-type GenesisReply struct {
-	Genesis *genesis.Genesis `json:"genesis"`
-}
-
-func (j *JSONRPCServer) Genesis(_ *http.Request, _ *struct{}, reply *GenesisReply) (err error) {
-	reply.Genesis = j.c.Genesis()
-	return nil
-}
-
-type TxArgs struct {
-	TxID ids.ID `json:"txId"`
-}
-
-type TxReply struct {
-	Timestamp int64           `json:"timestamp"`
-	Success   bool            `json:"success"`
-	Units     fees.Dimensions `json:"units"`
-	Fee       uint64          `json:"fee"`
-}
-
-func (j *JSONRPCServer) Tx(req *http.Request, args *TxArgs, reply *TxReply) error {
-	ctx, span := j.c.Tracer().Start(req.Context(), "Server.Tx")
-	defer span.End()
-
-	found, t, success, units, fee, err := j.c.GetTransaction(ctx, args.TxID)
-	if err != nil {
-		return err
-	}
-	if !found {
-		return ErrTxNotFound
-	}
-	reply.Timestamp = t
-	reply.Success = success
-	reply.Units = units
-	reply.Fee = fee
-	return nil
-}
