--- x/programs/cmd/simulator/vm/storage/errors.go
@@ -1,8 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package storage
-
-import "errors"
-
-var ErrInvalidBalance = errors.New("invalid balance")
