--- x/programs/cmd/simulator/vm/storage/storage.go
@@ -1,191 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package storage
-
-import (
-	"context"
-	"encoding/binary"
-	"errors"
-
-	"github.com/ava-labs/avalanchego/database"
-	"github.com/ava-labs/avalanchego/ids"
-
-	"github.com/ava-labs/hypersdk/consts"
-	"github.com/ava-labs/hypersdk/crypto/ed25519"
-	"github.com/ava-labs/hypersdk/fees"
-	"github.com/ava-labs/hypersdk/state"
-)
-
-const (
-	// metaDB
-	txPrefix = 0x0
-
-	// stateDB
-	keyPrefix = 0x0
-
-	programPrefix      = 0x1
-	heightPrefix       = 0x2
-	timestampPrefix    = 0x3
-	feePrefix          = 0x4
-	incomingWarpPrefix = 0x5
-	outgoingWarpPrefix = 0x6
-)
-
-var (
-	failureByte  = byte(0x0)
-	successByte  = byte(0x1)
-	heightKey    = []byte{heightPrefix}
-	timestampKey = []byte{timestampPrefix}
-	feeKey       = []byte{feePrefix}
-)
-
-const ProgramChunks uint16 = 1
-
-//
-// Program
-//
-
-func ProgramKey(id ids.ID) (k []byte) {
-	k = make([]byte, 1+consts.IDLen)
-	k[0] = programPrefix
-	copy(k[1:], id[:])
-	return
-}
-
-// [programID] -> [programBytes]
-func GetProgram(
-	ctx context.Context,
-	db state.Immutable,
-	programID ids.ID,
-) (
-	[]byte, // program bytes
-	error,
-) {
-	k := ProgramKey(programID)
-	return db.GetValue(ctx, k)
-}
-
-// setProgram stores [program] at [programID]
-func SetProgram(
-	ctx context.Context,
-	mu state.Mutable,
-	programID ids.ID,
-	program []byte,
-) error {
-	k := ProgramKey(programID)
-	return mu.Insert(ctx, k, program)
-}
-
-//
-// Keys
-//
-
-// gets the public key mapped to the given name.
-func GetPublicKey(ctx context.Context, db state.Immutable, name string) (ed25519.PublicKey, bool, error) {
-	k := make([]byte, 1+ed25519.PublicKeyLen)
-	k[0] = keyPrefix
-	copy(k[1:], []byte(name))
-	v, err := db.GetValue(ctx, k)
-	if errors.Is(err, database.ErrNotFound) {
-		return ed25519.EmptyPublicKey, false, nil
-	}
-	if err != nil {
-		return ed25519.EmptyPublicKey, false, err
-	}
-	return ed25519.PublicKey(v), true, nil
-}
-
-func SetKey(ctx context.Context, db state.Mutable, privateKey ed25519.PrivateKey, name string) error {
-	k := make([]byte, 1+ed25519.PublicKeyLen)
-	k[0] = keyPrefix
-	copy(k[1:], []byte(name))
-	return db.Insert(ctx, k, privateKey[:])
-}
-
-//
-// Transactions
-//
-
-func StoreTransaction(
-	_ context.Context,
-	db database.KeyValueWriter,
-	id ids.ID,
-	t int64,
-	success bool,
-	units fees.Dimensions,
-	fee uint64,
-) error {
-	k := txKey(id)
-	v := make([]byte, consts.Uint64Len+1+fees.DimensionsLen+consts.Uint64Len)
-	binary.BigEndian.PutUint64(v, uint64(t))
-	if success {
-		v[consts.Uint64Len] = successByte
-	} else {
-		v[consts.Uint64Len] = failureByte
-	}
-	copy(v[consts.Uint64Len+1:], units.Bytes())
-	binary.BigEndian.PutUint64(v[consts.Uint64Len+1+fees.DimensionsLen:], fee)
-	return db.Put(k, v)
-}
-
-func GetTransaction(
-	_ context.Context,
-	db database.KeyValueReader,
-	id ids.ID,
-) (bool, int64, bool, fees.Dimensions, uint64, error) {
-	k := txKey(id)
-	v, err := db.Get(k)
-	if errors.Is(err, database.ErrNotFound) {
-		return false, 0, false, fees.Dimensions{}, 0, nil
-	}
-	if err != nil {
-		return false, 0, false, fees.Dimensions{}, 0, err
-	}
-	t := int64(binary.BigEndian.Uint64(v))
-	success := true
-	if v[consts.Uint64Len] == failureByte {
-		success = false
-	}
-	d, err := fees.UnpackDimensions(v[consts.Uint64Len+1 : consts.Uint64Len+1+fees.DimensionsLen])
-	if err != nil {
-		return false, 0, false, fees.Dimensions{}, 0, err
-	}
-	fee := binary.BigEndian.Uint64(v[consts.Uint64Len+1+fees.DimensionsLen:])
-	return true, t, success, d, fee, nil
-}
-
-// [txPrefix] + [txID]
-func txKey(id ids.ID) (k []byte) {
-	k = make([]byte, 1+consts.IDLen)
-	k[0] = txPrefix
-	copy(k[1:], id[:])
-	return
-}
-
-func HeightKey() (k []byte) {
-	return heightKey
-}
-
-func TimestampKey() (k []byte) {
-	return timestampKey
-}
-
-func FeeKey() (k []byte) {
-	return feeKey
-}
-
-func IncomingWarpKeyPrefix(sourceChainID ids.ID, msgID ids.ID) (k []byte) {
-	k = make([]byte, 1+consts.IDLen*2)
-	k[0] = incomingWarpPrefix
-	copy(k[1:], sourceChainID[:])
-	copy(k[1+consts.IDLen:], msgID[:])
-	return k
-}
-
-func OutgoingWarpKeyPrefix(txID ids.ID) (k []byte) {
-	k = make([]byte, 1+consts.IDLen)
-	k[0] = outgoingWarpPrefix
-	copy(k[1:], txID[:])
-	return k
-}
