--- x/programs/cmd/simulator/vm/version/version.go
@@ -1,12 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package version
-
-import "github.com/ava-labs/avalanchego/version"
-
-var Version = &version.Semantic{
-	Major: 0,
-	Minor: 0,
-	Patch: 1,
-}
